+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.1.1 (Unreleased)
+  * Fix dummy-so generation to use correct syntax for ARM with `--dummy-so=yes`
 
 # 2.1.0
   * `--asm` option now prints the assembly for each module of an IR separately